fix gcc-13 -Werror=useless-cast violations across 4 files
GCC 13's -Werror=useless-cast (enabled in cmake/SouxmarWarnings.cmake
for the GNU toolchain) flags six explicit casts where the source
type equals the target type on every platform souxmar supports:

  include/souxmar-c/buffer.h:85  (uint32_t)1u << 0
  include/souxmar-c/buffer.h:86  (uint32_t)1u << 1
  include/souxmar-c/mesh.h:128   (int32_t)-1
  src/core/c_abi_buffer.cpp:99   static_cast<std::uintptr_t>(kAlignment - 1)
  src/core/mesh.cpp:94           static_cast<std::uint64_t>(vec.size())
  src/core/mesh.cpp:130          static_cast<std::size_t>(end - begin)

On every supported platform/triplet (x86_64 + arm64 on linux,
macos, windows):
  - unsigned int   == uint32_t  (32-bit)
  - int            == int32_t   (32-bit)
  - std::size_t    == uintptr_t == uint64_t  (64-bit)

The casts therefore produce token-identical output and were
flagged correctly as useless. Dropping them preserves the exact
runtime value, the exact bit pattern, and the exact ABI surface
contract — the macros still expand to a literal of type uint32_t /
int32_t respectively, the size_t conversions still produce size_t.

Frozen-header note: include/souxmar-c/buffer.h and mesh.h are
under ADR-0008's frozen-ABI rule. The edits here change only the
token sequence of the macros' replacement-list; they do not
change the macro's value, the macro's expanded type, the public
type of any function declaration, or anything a linker would
observe. No symbol added, removed, renamed, or re-typed. The
on-disk pipeline format and the agent tool contract are not
touched.

This unblocks the gcc-13 CI build that started actually compiling
after PR #8's vcpkg + gcc-13-install fixes landed.

Ratchet: bug-fix (ADR-0008) — drop useless casts gcc-13 flags

+++ include/souxmar-c/buffer.h
@@ -82,8 +82,8 @@ size_t souxmar_buffer_alignment(void);
  * souxmar_status_t around the operation.
  */
 
-#define SOUXMAR_BUFFER_FLAG_READONLY ((uint32_t)1u << 0)
-#define SOUXMAR_BUFFER_FLAG_CREATE ((uint32_t)1u << 1)
+#define SOUXMAR_BUFFER_FLAG_READONLY (1u << 0)
+#define SOUXMAR_BUFFER_FLAG_CREATE (1u << 1)
 
 souxmar_buffer_t* souxmar_buffer_new_mmap(const char* path, size_t size_bytes, uint32_t flags);
